Tests: Fix PHPStan and coding standard violations
- Construct single-consumer query services directly (inlined out of the
  compiled container, so container->get() fails PHPStan's symfony extension)
- Narrow array-shape offsets with assertArrayHasKey before access
- Move MessageBusSpy to tests/TestDouble (PSR1 one-class-per-file)
