+++ tests/TestDouble/MessageBusSpy.php
@@ -1,0 +1,21 @@
+<?php
+
+declare(strict_types=1);
+
+namespace SpeedPuzzling\Web\Tests\TestDouble;
+
+use Symfony\Component\Messenger\Envelope;
+use Symfony\Component\Messenger\MessageBusInterface;
+
+final class MessageBusSpy implements MessageBusInterface
+{
+    /** @var list<object> */
+    public array $dispatched = [];
+
+    public function dispatch(object $message, array $stamps = []): Envelope
+    {
+        $this->dispatched[] = $message;
+
+        return new Envelope($message);
+    }
+}
